Solve integration issues - not done
